Invert relationship of Player and PlayerControls
This makes more sense. The logic component owns the state and passes props to the
UI.
